modularizacao do codigo + pydantic
