+++ app/schema.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class ProdutosSchema(BaseModel):
+    id: int
+    nome: str 
+    descricao: str 
+    preco: float
