+++ tests/test_api.py
@@ -1,6 +1,6 @@
 import pytest 
 from fastapi.testclient import TestClient
-from main import app
+from app.main import app
 
 
 client = TestClient(app)
